requirements.txt
requests>=2.31.0
pillow>=10.2.0
# 其他需要的依赖包 